confirm package/theme uninstall

+++ app/system/modules/package/app/components/extensions.js
@@ -59,11 +59,16 @@ module.exports = {
         },
 
         uninstall: function (pkg) {
-            this.uninstallPackage(pkg, this.packages).success(function (data) {
-                UIkit.notify(this.$trans('"%title%" uninstalled.', {title: pkg.title}));
-            }).error(function (data) {
-                UIkit.notify(data, 'danger');
-            });
+
+            UIkit.modal.confirm(this.$trans('Are you sure?'), function() {
+
+                this.uninstallPackage(pkg, this.packages).success(function (data) {
+                    UIkit.notify(this.$trans('"%title%" uninstalled.', {title: pkg.title}));
+                }).error(function (data) {
+                    UIkit.notify(data, 'danger');
+                });
+                
+            }.bind(this));
         }
 
     }
